+++ src/test/run.ts
@@ -3,6 +3,7 @@ import "./normalize.test";
 import "./sync.test";
 import "./query-demo.test";
 import "./api.test";
+import "./protocol.test";
 import { run } from "./testKit";
 
 await run();